Add: プレイヤーステータスを別コンポーネントに移した
