+++ frontend/src/containers/MyPages.jsx
@@ -6,7 +6,7 @@ import styled from 'styled-components';
 import { Header } from '../components/Headers/Header.jsx';
 import { FakeHeader } from '../components/Headers/FakeHeader.jsx';
 import { SessionFlashMessage } from '../components/FlashMessages/SessionFlashMessage.jsx';
-import { StatusExperienceBox } from '../components/Games/StatusExperienceBox.jsx';
+import { UserStatus } from '../components/Games/UserStatus.jsx';
 import { StudyHeatMap } from '../components/Games/StudyHeatMap.jsx';
 import { CorrectPercentGraph } from '../components/Games/CorrectPercentGraph.jsx';
 import { ChangeGraphBox } from '../components/Games/ChangeGraphBox.jsx';
@@ -14,7 +14,6 @@ import { DescriptionWrapper } from '../components/shared_style.js';
 import { GameContent } from '../components/GameContents/GameContent.jsx';
  
 // Images
-import TemporaryUserImage from '../images/temporary_user_image.png';
 import ElementaryGameContentImage from '../images/elementary_game_content.png';
 import IntermediateGameContentImage from '../images/intermediate_game_content.png';
 
@@ -63,85 +62,6 @@ const MainSecondWrapper = styled.div`
 const SecondWrapper = styled.div`
   background-color: ${COLORS.SUB};
   padding-top: 80px;
-`;
-
-// ステータスのラッパー
-const StatusWrapper = styled.div`
-  width: 700px;
-  height: 260px;
-  border-radius: 3px;
-  background-color: ${COLORS.WHITE};
-  box-shadow: 0 0px 20px rgba(0,0,0,0.2);
-  display: flex
-`;
-
-// イメージのラッパー
-const ImageWrapper = styled.div`
-  width: 220px;
-  height: 220px;
-  background-color: ${COLORS.WHITE};
-  border-radius: 3px;
-  align-self: center;
-  padding: 16px;
-`;
-
-// カスタムイメージ
-// imageタグだと縦横比を維持できない
-// background-imageだと縦横比を維持できる
-const CustomImage = styled.div`
-  background-image: url(${TemporaryUserImage});
-  width: 220px;
-  height: 220px;
-  border-radius: 50%;
-  background-size: cover;
-  background-position: center center;
-  border: 3px solid silver;
-  box-shadow: inset 1px 1px 3px 0 rgba(0, 0, 0, 0.8), 1px 1px 0 0 rgba(255, 255, 255, 0.12);
-`;
-
-const TableWrapper = styled.div`
-  align-self: center;
-`;
-
-const CustomTable = styled.table`
-  width: 430px;
-  border-collapse: collapse;
-  color: ${COLORS.BLACK};
-  font-family: YuGothic;
-  font-weight: normal;
-  font-size: 18px;
-  margin: 0 auto;
-  border: none;
-`;
-
-const CustomTd = styled.td`
-  padding: 10px 40px; 
-  border: none;
-  text-align: right;
-  border-bottom:solid 1px silver;
-`;
-
-const NameTd = styled(CustomTd)`
-  font-family: YuGothic;
-  font-style: normal;
-  font-weight: bold;
-  font-size: 32px;
-  line-height: 40px;
-  text-align: left;
-  color: ${COLORS.BLACK};
-`;
-
-const MetaTd = styled(CustomTd)`
-  padding: 10px 40px; 
-  border: none;
-  text-align: left;
-  border-bottom:solid 1px silver;
-  font-weight: bold;
-`;
-
-const ExpTd = styled(CustomTd)`
-  text-align: left;
-  border: none;
 `;
 
 const StudyHeatMapWrapper = styled.div`
@@ -317,37 +237,14 @@ export const MyPages = () => {
       </FakeBlock>
       <MainWrapper>
         <MainFirstWrapper>
-          <StatusWrapper>
-            <ImageWrapper>
-              <CustomImage />
-            </ImageWrapper>
-            <TableWrapper>
-              <CustomTable>
-                <tbody>
-                  <tr>
-                    <NameTd colSpan={2}>{user.name}</NameTd> 
-                  </tr>
-                  <tr>
-                    <MetaTd>ランク</MetaTd> 
-                    <CustomTd>{user.rank}</CustomTd>
-                  </tr>
-                  <tr>
-                    <MetaTd>称号</MetaTd> 
-                    <CustomTd>{user.active_title}</CustomTd>
-                  </tr>
-                  <tr>
-                    <ExpTd colSpan={2}>
-                      <StatusExperienceBox 
-                        temporary_experience={user.temporary_experience}
-                        total_experience={user.total_experience}
-                        maximum_experience_per_rank={user.maximum_experience_per_rank}
-                      />
-                    </ExpTd>
-                  </tr>
-                </tbody>
-              </CustomTable>
-            </TableWrapper>
-          </StatusWrapper>
+          <UserStatus
+            name={user.name}
+            rank={user.rank}
+            active_title={user.active_title}
+            temporary_experience={user.temporary_experience}
+            total_experience={user.total_experience}
+            maximum_experience_per_rank={user.maximum_experience_per_rank}
+          />
           <StudyHeatMapWrapper>
             <StudyHeatMap 
               game_frequencies_per_day={myPageState.game_frequencies_per_day}
